% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin_dataname.R
\name{begin_dataname}
\alias{begin_dataname}
\title{Loading and Naming a Dataset}
\usage{
begin_dataname(path)
}
\arguments{
\item{path}{The path to the csv file to be loaded}
}
\value{
A loaded and assigned dataset
}
\description{
Loading and Naming a Dataset
}
\examples{
begin_dataname("path/to_file.csv")

}
