% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin_vizguide.R
\name{begin_vizguide}
\alias{begin_vizguide}
\title{Providing Links for Data Visualizations}
\usage{
begin_vizguide()
}
\value{
A List of links and text
}
\description{
Providing Links for Data Visualizations
}
\examples{
begin_vizguide()

}
