% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin_datatype.R
\name{begin_typecheck}
\alias{begin_typecheck}
\title{Determine the Datatype}
\usage{
begin_typecheck(x)
}
\arguments{
\item{x}{One value of which the data typ is to be determined (factor, charackter, numeric)}
}
\value{
A character naming the typ of data
}
\description{
Determine the Datatype
}
\examples{
begin_typecheck(123)

}
